CHANGED: notifications scheme
This commit goes with notifications scheme changed.

Also with this commit:

SPLIT: Test doubles.
